Added "sleeping" behavior
The images will fade out if the mouse isn’t moving after a little
while. Numbers will likely have to be updated.

Right now, it takes ample time to fade in, is patient while nothing is
happening, and then fades out quickly. Hopefully this encourages a
learning loop so people realize they have to keep moving to see the
images.